ajax editing posts

+++ app/Http/Controllers/PostController.php
@@ -22,9 +22,23 @@ class PostController extends Controller
         return redirect()->route('home')->with(['message' => $message]);
     }
 
+    public function postEditPost(Request $request)
+    {
+        $this->validate($request, [
+           'body' => 'required'
+        ]);
+        $post = Post::find($request['postId']);
+        $post->body = $request['body'];
+        $post->update();
+        return response()->json(['new_body' => $post->body], 200);
+    }
+
     public function getDeletePost($post_id)
     {
         $post = Post::where('id', $post_id)->first();
+        if (Auth::user() != $post->user) {
+            return redirect()->back();
+        }
         $post->delete();
         return redirect()->route('home')->with(['message' => 'Successfully deleted!']);
     }
